requirements.txt

#### **requirements.txt**

pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
web3>=5.24.0
aiohttp>=3.8.0
PyYAML>=6.0
torch>=1.10.0
torch_geometric>=2.0.0
joblib>=1.1.0